Refactored upload methods
