+++ src/upload/errors.rs
@@ -14,12 +14,6 @@ pub enum UploadError {
     #[error("Invalid number of files {0}, there should be an even number of files")]
     InvalidNumberOfFiles(usize),
 
-    #[error("No Bundlr balance found for address: {0}, check Bundlr cluster and address balance")]
-    NoBundlrBalance(String),
-
-    #[error("Invalid Bundlr cluster: {0} Use 'devnet' or 'mainnet'")]
-    InvalidBundlrCluster(String),
-
     #[error("{0}")]
     Incomplete(String),
