+++ src/upload/methods/mod.rs
@@ -1,0 +1,5 @@
+pub mod aws;
+pub mod bundlr;
+
+pub use aws::*;
+pub use bundlr::*;
